Addressing review comments
- Renaming the packages for model classes
- Refactoring test cases

Part of #610
