+++ src/main/java/com/sap/oss/phosphor/fosstars/data/artifact/VulnerabilitiesFromOwaspDependencyCheck.java
@@ -8,6 +8,12 @@ import com.sap.oss.phosphor.fosstars.data.DataProvider;
 import com.sap.oss.phosphor.fosstars.data.NoValueCache;
 import com.sap.oss.phosphor.fosstars.data.UserCallback;
 import com.sap.oss.phosphor.fosstars.data.ValueCache;
+import com.sap.oss.phosphor.fosstars.data.owasp.model.Dependency;
+import com.sap.oss.phosphor.fosstars.data.owasp.model.OwaspDependencyCheckEntry;
+import com.sap.oss.phosphor.fosstars.data.owasp.model.OwaspDependencyCheckReference;
+import com.sap.oss.phosphor.fosstars.data.owasp.model.OwaspDependencyCheckVuln;
+import com.sap.oss.phosphor.fosstars.data.owasp.model.Software;
+import com.sap.oss.phosphor.fosstars.data.owasp.model.VulnerableSoftware;
 import com.sap.oss.phosphor.fosstars.model.Feature;
 import com.sap.oss.phosphor.fosstars.model.Subject;
 import com.sap.oss.phosphor.fosstars.model.ValueSet;
@@ -19,12 +25,6 @@ import com.sap.oss.phosphor.fosstars.model.value.Vulnerabilities;
 import com.sap.oss.phosphor.fosstars.model.value.Vulnerability;
 import com.sap.oss.phosphor.fosstars.model.value.Vulnerability.Builder;
 import com.sap.oss.phosphor.fosstars.model.value.Vulnerability.Resolution;
-import com.sap.oss.phosphor.fosstars.owasp.data.Dependency;
-import com.sap.oss.phosphor.fosstars.owasp.data.OwaspDependencyCheckEntry;
-import com.sap.oss.phosphor.fosstars.owasp.data.OwaspDependencyCheckReference;
-import com.sap.oss.phosphor.fosstars.owasp.data.OwaspDependencyCheckVuln;
-import com.sap.oss.phosphor.fosstars.owasp.data.Software;
-import com.sap.oss.phosphor.fosstars.owasp.data.VulnerableSoftware;
 import com.sap.oss.phosphor.fosstars.util.Json;
 import java.io.File;
 import java.io.FileOutputStream;
