+++ src/main/java/com/sap/oss/phosphor/fosstars/data/owasp/model/package-info.java
@@ -2,4 +2,4 @@
  * The package contains POJOs which was manually generated based on a sample OWASP Dependency-Check
  * result.
  */
-package com.sap.oss.phosphor.fosstars.owasp.data;
+package com.sap.oss.phosphor.fosstars.data.owasp.model;
