+++ src/test/resources/com/sap/oss/phosphor/fosstars/model/rating/oss/OssArtifactSecurityRatingTestVectors.yml
@@ -206,7 +206,54 @@ defaults:
       type: "VulnerabilitiesFeature"
       name: "Info about vulnerabilities in open-source project"
     vulnerabilities:
-      entries: []
+      entries: 
+        - id: "VULN-123-1"
+          cvss:
+            version: "V3"
+            value: 9.0
+          references: []
+          resolution: "PATCHED"
+          introduced: "2019-01-01"
+          fixed: "2019-01-03"
+          published: "2019-01-03"
+          vulnerableVersions:
+            - versionStart: 1.0.0
+              versionEnd: 1.1.0
+        - id: "VULN-123-2"
+          cvss:
+            version: "V3"
+            value: 9.0
+          references: []
+          resolution: "PATCHED"
+          introduced: "2019-01-01"
+          fixed: "2019-01-03"
+          published: "2019-01-03"
+          vulnerableVersions:
+            - versionStart: 1.0.0
+              versionEnd: 2.1.0
+        - id: "VULN-124-2"
+          cvss:
+            version: "V3"
+            value: 7.0
+          references: []
+          resolution: "PATCHED"
+          introduced: "2018-11-28"
+          fixed: "2018-12-02"
+          published: "2019-01-03"
+          vulnerableVersions:
+            - versionStart: 1.1.0
+              versionEnd: 1.2.0
+        - id: "VULN-123-3"
+          cvss:
+            version: "V3"
+            value: 9.0
+          references: []
+          resolution: "UNPATCHED"
+          introduced: "2019-01-01"
+          published: "2019-01-03"
+          vulnerableVersions:
+            - versionStart: 1.0.0
+              versionEnd: 1.1.0
 
 # test vectors
 elements:
@@ -352,10 +399,10 @@ elements:
             date: "2020-01-02"
     expectedScore:
       type: "DoubleInterval"
-      from: 8.0
+      from: 7.0
       openLeft: false
       negativeInfinity: false
-      to: 9.0
+      to: 8.0
       openRight: false
       positiveInfinity: false
     expectedLabel:
@@ -372,7 +419,7 @@ elements:
           name: "Info about vulnerabilities in open-source artifact"
         vulnerabilities:
           entries:
-            - id: "VULN-123-1"
+            - id: "VULN-123-2"
               cvss:
                 version: "V3"
                 value: 9.0
@@ -426,7 +473,7 @@ elements:
           name: "Info about vulnerabilities in open-source artifact"
         vulnerabilities:
           entries:
-            - id: "VULN-123-1"
+            - id: "VULN-123-2"
               cvss:
                 version: "V3"
                 value: 9.0
@@ -484,11 +531,6 @@ elements:
         feature:
           type: "BooleanFeature"
           name: "If an open-source project has a security team"
-        flag: false
-      - type: "BooleanValue"
-        feature:
-          type: "BooleanFeature"
-          name: "If an open-source project belongs to Apache Foundation"
         flag: false
       - type: "LgtmGradeValue"
         feature:
@@ -583,7 +625,7 @@ elements:
           name: "Info about vulnerabilities in open-source artifact"
         vulnerabilities:
           entries:
-            - id: "VULN-123-1"
+            - id: "VULN-123-3"
               cvss:
                 version: "V3"
                 value: 9.0
